requirements.txt
anyascii==0.3.1
asgiref==3.5.2
Babel==2.10.3
beautifulsoup4==4.9.3
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-crum==0.7.9
django-filter==21.1
django-modelcluster==6.0
django-money==3.0.0
django-permissionedforms==0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
mysqlclient==2.1.1
openpyxl==3.0.10
Pillow==9.2.0
py-moneyed==2.0
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
typing_extensions==4.3.0
urllib3==1.26.10
wagtail==3.0.1
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
